Stub TestDebug JS implementation so we're not failing the build until it's implemented.
