+++ test/js/src/main/scala/zio/test/TestDebug.scala
@@ -1,0 +1,35 @@
+package zio.test
+
+import zio.{Ref, ZIO}
+
+private[test] object TestDebug {
+  def printEmergency(executionEvent: ExecutionEvent, lock: Ref.Synchronized[Unit]) =
+    executionEvent match {
+      case t@ExecutionEvent.TestStarted(
+      labelsReversed,
+      annotations,
+      ancestors,
+      id,
+      fullyQualifiedName
+      ) =>
+        write(s"${t.labels.mkString(" - ")} STARTED\n", true, lock)
+
+      case t@ExecutionEvent.Test(labelsReversed, test, annotations, ancestors, duration, id, fullyQualifiedName) =>
+        removeLine(t.labels.mkString(" - ") + " STARTED")
+
+      case ExecutionEvent.SectionStart(labelsReversed, id, ancestors) => ZIO.unit
+      case ExecutionEvent.SectionEnd(labelsReversed, id, ancestors) => ZIO.unit
+      case ExecutionEvent.TopLevelFlush(id) =>
+        ZIO.unit
+      case ExecutionEvent.RuntimeFailure(id, labelsReversed, failure, ancestors) =>
+        ZIO.unit
+    }
+
+  // TODO Implement this with appropriate JS filesystem APIs after JVM version is finalized
+  def write(content: => String, append: Boolean, lock: Ref.Synchronized[Unit]): ZIO[Any, Nothing, Unit] =
+    ZIO.unit
+
+  private def removeLine(searchString: String) =
+    ZIO.unit
+
+}
